Moved common settings activity logic to an abstract class in library module.
